Fixes for pr
